CAPIO server code refactoring
This commit applies a huge refactor to the CAPIO codebase. It creates
three different modules:

  - A `common` module, which contains all the utility functions and data
    structures shared among all the other CAPIO modules;
  - A `posix` module, which contains the logic related to the
    `libcapio_posix.so` wrapper;
  - A `server` module, which contains the logic related to the
    `capio_server` binary.

In addition, this commit solves the recursive logging problem by
substituting the `vfprintf` function (which calls `SYS_write`) with the
`vsnprintf` function, which writes the logs into a string buffer that is
then flushed to `stderr` through a `syscall_no_intercept(SYS_write)`
call.