Improved json config parser
This commit slightly improves the code for json.hpp.
The functions to
 - parse input file section
 - parse output file sections
needs to be coded as are now missing.
This commit also introduces log messages so that on screen the status of
code parsing is shown.
The section for parsing permanent and streaming sections still needs to
be improved