Fixes for stat
This commit introduces tests for stat syscall.
It also fixes a problem in which when stat is called for a file that is
not yet present in capio, only the size was returned to the client app,
when also the is_dir must be sent. now bth parameters are returned as
-1;