Improve unit test CI
This commit adds a timeout of 2 minutes to the unit test CI step in
order to avoid useless waiting when a test deadlocks. It also ensures
that the server log is correctly printed whenever the unit test step
fails, for debugging purposes.

This commit also moves the `CAPIO_DIR` to the `build/tests` folder, in
order to simplify the testing scenario. Indeed the `CAPIO_DIR` should be
empty prior to the application start.