Improved java compatibility
This commits fix a wrong behavior for capio_posix_realpath, in which
instead of returning an empty string if path is neither capio or posix,
the client app would crash. Thos was wrong as one might want to try to
open a file to see if it exists, and if fails then file do not exists.
now instead of calling ERR_EXIT(), an empty string is returned and a log
message added. It is the up to the function that uses
capio_posix_realpath to check whether the string has length > 0.

This commit also fixes an error on SYS_getcwd and also adds
compatibility to java and python

Tests for SYS_getcwd are added