Fix `capio_posix_path` and `getcwd` behaviour (#19)
This commits fix a wrong behavior for `capio_posix_realpath`,
in which instead of returning an empty string if a file does not
exist, the client app would crash. This is a wrong behaviour:
one might want to try to open a file just to see if it actually exists.
With this commit, an empty `std::string` object is returned instead
of calling `ERR_EXIT`. It is now up to the caller of
`capio_posix_realpath` to check whether the string has length > 0.

This commit also fixes th behaviour of the `getcwd` syscall,
which incorrectly converted a `char *` pointer into a `std::string`
type. The `getcwd` unit test has also been restored.