Improve the `syscallnames.h` file generation
THis commit moves the `syscallnames.h` auto-generated file from the
source folders (managed by git) to the binary folders. This fix avoids
problems with spurious uncommitted files whenever developers want to
commit their code, and it avoids the risk to erroneously commit the
`syscallnames.h` files in the git repository.

In addition, this commit improves the way the `unistd_${ARCH}.h` file is
retrieved from the OS, making the logic more general and portable.