Fix error handling in `stat` syscall family (#48)
This commit fixes error handling logic in the `stat` family of system
calls, ensuring that the correct `errno` value is returned when
something goes wrong. Before this commit, this part was flawed and
sometimes CAPIO returned succesfully even in presence of errors, e.g.,
when stat was executed on a nonexistent file.