Improve logger for multiple threads (#49)
This commit adds the `tid` marker also on the header and footer messages
of POSIX syscalls and server requests in the CAPIO `Logger`, to improve
readability when multiple threads are running.

This commit also stop header and footer from showing up when a new
thread is called through the `hook_clone_child` function, as they
reported misleading syscall numbers that could cause confusion. This is
achieved by introducing a new `SyscallLoggingSuspender` class that
adheres to the C++ RAII principle.