Fix `hook_clone_child` SIGSEGV in glibc (#50)
Freeing memory inside the `hook_clone_child` function is dangerous, as
the `free` glibc function tries to access the `__ctype_b_loc` table but
sometimes it could not be initialized yet, causing a SIGSEGV error.
THis commit moves the `tids.erease()` call in the `exit_group_handler`
function, when a process is deleted. Plus, this commit ensures that all
threads waiting on the `clone_cv` are notified when a child is
initialized, preventing deadlocks.