Reformat constants and add CAPIO forbidden paths
This commit adds support for forbidden paths for the `CAPIO_DIR`, which
result in early exit of CAPIO logic without memory allocations. This is
crucial to support those (rare) cases when `malloc` and other primitive
libc functions read the file system.

In addition, this commit uniforms the CAPIO constants names and reorders
them in the `constants.hpp` file to improve code readability.