Fixed capio server logger on multinode
PReviusly capiio did use the rank to differentiate between different
server ranks, before the backend was not available, resultin in all
server logging to capio_server_0.log. now hostname is used instead of
rank.

Also the logfile is not created when logging is not compiled into capio.
this was de to a bad management of logging capabilities into the server
argument parsing function.