Server bugfixes (#42)
This commit improves the general readability of the CAPIO server
backend and fixes the logger for multi-node support. Previously
CAPIO was using the MPI rank to differentiate between server
instances, resulting in all server instances logging to the
`capio_server_0.log` file when the MPI  backend is not selected.
Now it uses the more general `hostname` in place of the rank.

Also, the logfile is no more created when logging is not compiled
into CAPIO. This was happening because of a bad management
of logging capabilities into the server's argument parser.