Fixed remote read
This commit fixes the remote read procedure as well as adding more
logging to capio. in particular in
server/comm/remote_listener.hpp::remote_listener_remote_sending()
function, file_shm allways resulted in being not initialized. which
caused an error in mpi recive. This was due to missing else branch on
checking for file size.