Rewrite logic of `load_file_location` function (#54)
Before this commit, this function was named `check_file_location` and
its behaviour was unclear. This function does not only verify if a file
exists on any node: it also loads its location to the current node cache
if possible. The function returns `false` only when the file is not
present on any CAPIO node.

There were two different versions of this function. This commit merges
them into a single one, as the innermost one was never called directly.

Finally, this commit reduces the calls to the `load_file_location`
function. Since this function is quite heavy, especially in large
deplyoments, it is better to call it only when strictly necessary.