Fixed remote stat handler behaviour when streaming
When a stat is requested on a capio file that has not been yet created,
but will be (as the path is matched by a glob), the stat now lookg both
on files_metadata to see whether the file has alreasy been created, and
on match globs to see whether the file will be created in the future. in
the latter case, then capio will initialize the file in advance and wait
for it to be streamable.