Delete `CapioFile` copy constructors (#63)
This commit deletes the `CapioFile` copy constructors to prepare it for
containing `std::mutex` objects. To do that, it converts the
`init_capio_file` function to a method of the `CapioFile` class, in
order to avoid spurious copy.

This commit introduces also other small fixes and refactors, like
deleting further unwanted copy constructors.