Refactor synchronization protocol for `read` (#61)
This commit rewrites the synchronization protocol for the 
`read` syscall. In particular, it replaces semaphores wit
`std::mutex` and `std::condition_variable` objects and replaces
explicit awake methods with the condition variable `wait` and
`notify` mechanisms. This is in line with the synchronization
protocol refactoring implemented for the `stat` syscall in #64.