Removed references to rank in capio server
This commit changes the behaviour of server by removing references to
rank, which is only used by MPI and by using hostnames. In the backend a
private structure is used to convert from hostname to rank and vice
versa. In the future FQD might be wanted.

Also removed global data structures to convert from hostnames to ranks
and vice versa