Fix `getcwd` return value
The `sys_getcwd` system call should return the length of the path stored
in the `buf` argument. Prior to this commit, the CAPIO `getcwd_handler`
was not returning any explicit result in case of success, leaving the
`result` variable point to an unitialized memory region. This behaviour
was the cause of several issues. This commit fixes the `getcwd_handler`,
implementing the standard POSIX behaviour.