Clear CAPIO path cache after `unlink`
The `delete_capio_path` method removes a path from the local CAPIO POSIX
cache. However, before this commit, it was invoked just after `rmdir`,
leaving leftovers whenever an `unlink` system call was invoked. This
commit fixes this behaviour.

In addition, this commit changes the internal logic of the
`delete_capio_path` method to avoid segmentation faults when a `fd`
object is removed from the set during the loop iterations.