Run unit tests in Docker (#82)
Before this commit the CI/CD pipeline only ran few example command in
the built Docker container to check that nothing critical happened to
the build process. This commit runs the entire suite of CAPIO unit tests
in the built Docker container to better check that everything is
working. This commit also enables the possibility to copy test suites in
the released Docker container.