Reverted files locations (#70)
This commit restores the original architecture of the `files_location`
database, which was using a different file for each of the involved
compute nodes. This approach avoids race conditions on parallel file
systems that do not support distributed `flock` synchronization. Plus,
it reduces the read/write contention on each of the files, improving
performance.

---------

Co-authored-by: Marco Edoardo Santimaria <santimaria@di.unito.it>
Co-authored-by: GlassOfWhiskey <iacopo.c92@gmail.com>