Store independent file offsets in client and server
Before this commit, a file's offset was stored into a shared memory
location to be accessed and update by both CAPIO client and server.
Hpwever, this implementation is tricky for implementing read prefetching
and write buffering on the client side, because in that setting the two
offsets should be different. This ocmmit separates the two values,
ensuring that both are kept updated in a coherent way.