Added option for injection of debug code (#88)
This commit adds the option to inject lambda functions to
debug at runtime the CAPIO code. The new `#define DBG` 
macro requires a lambda function in the form of `[](){}()`
so that parameters can be passed to it without the need of
variadic functions. This macro includes the lambda function
only when compiling in Debug and lefts out the lambda when
compiling in `Release` mode.