Unified `SPSCQueue`, `CircularBuffer`, and `Semaphore` improvements (#74)
Ths commit refactors the shared memory management in CAPIO. In particular
it unifies the `SPSCQueue` and `CircularBuffer` clases by using c++ templates
metaprogramming. This was possible as the difference between the two classes
are minimal. 

This commit also fixes issues that were preventing communication to
occur when using the `CAPIO_WORKFLOW_NAME` environment variable. 

C-style raw`sem_t` objects have been wrapped around C++ classes
to simplify the usage of semaphores. Two classes called `Semaphore`
and `NamedSemaphore` have been created. The difference is that the
`NamedSemaphore` creates a semaphore on shared memory, supporting
multi-process synchronization.

Finaly a `--continue-on-error` flag to the CAPIO server has been added.
This flag enables the CAPIO server to treat errors that would normally make
it crash as simple warnings. This flag is available only if the CAPIO logger is
enabled. Indeed, this behaviour is too dangerous to be allowed in a production
environment.