Store independent file offsets in client and server (#87)
Before this commit, a file's offset was stored into a shared memory location to be accessed and update by both CAPIO client and server. However, this implementation is tricky for implementing read prefetching and write buffering on the client side, because in that setting the two offsets should be different. This commit separates the two values, ensuring that both are kept updated in a coherent way.