Merge pull request #132 from High-Performance-IO/hotfix
This commit improves the logging status when failing to allocate
shared memory segments, and when ERR_EXIT is called.
It also adds a mechanism to handle applications that are highly
parallel, for which all threads (except the main) perform ONLY computation.
In this case, to avoid the exhaustion of file descriptors, CAPIO will not
intercept and handle child threads. To allow for this behaviour, export the
following variable ''CAPIO_IGNORE_CHILD_THREADS=YES''