Reduces SHM element allocation
This commit introduces a new category of response queue for requests
which uses two, instead of five, shared memory object for each instanced
element. The size is fixed to 64 bit, and can only be used to reply to
the posix application, offsets or numerical values that fits within 64
bits of space.