Added checks on committement of file...
This commit changes the behaviour of posix component and server
component of listdir. When called, it no longer directly asks if a file
is committed, but asks for the commit token path. It then caches it, and
periodically checks if said token exists. as soon as it does, it returns
NULL, as specified by the POSIX semantic for listdir||()