removed CAPIO_DIR from CAPIO_SERVER (#149)
This commit simplifies the CAPIO server component, by removing the requirement to have the CAPIO_DIR directory provided as an environment variable. 

This in turn allows different posix applications to execute with different CAPIO_DIR, but with the same CAPIO_WORKFLOW_NAME.

Finally when the capio_server parser encounters a relative path, this path is resolved to an absolute one, with prefix the current working directory of the capio server. This prefix can be changed with the command line option  --resolve-capiocl-to 

---------

Co-authored-by: Marco Edoardo Santimaria <marcoedoardo.santimaria@unito.it>