Added draft implementation of control plane (#161)
This commit adds the base mechanism to implement a multicast control plane with basic messages (CREATE and DELETE). 
It also adds the environment to develop locally the distributed capio_server trough the usage of docker. 

The implementation of the control plane is pending a major refactor of the capio_server structure

---------

Co-authored-by: Marco Edoardo Santimaria <marcoedoardo.santimaria@unito.it>