Refactor (#162)
This commits proceeds to refactor the capio project. It reorganizes the structure of the project, and refactors the capio-server component. In particular it splits definition of classes and implementation to avoid recursive
include issues that were rising when defining the capio_server backend. It also updates the tests to work with the new structure of the project, as well as the Dockerfile.
---------

Co-authored-by: Marco Edoardo Santimaria <marcoedoardo.santimaria@unito.it>