CAPIO-CL exlude section support (#167)
This commit introduces support for the exclude section of the CAPIO-CL file
Now when a file that should be excluded is opened, the library skips the
opening of the file, and the registration of the file descriptor as a
CAPIO file description. When the enumeration of a directory occurs, 
if the directory is marked as exclude, a commit token for that directory is created to 
ensure that the real readdir method is used.

---------

Co-authored-by: Marco Edoardo Santimaria <marcoedoardo.santimaria@unito.it>