Backend integration (#163)
This commits adds the backend to allow in-memory IO operation when running in multinode configurations. The developed backend wraps around the MTCL library, which itself is wrapped around by the CapioRemoteFile class. As of right now, only the create home-node-policy is supported. With this commit, the backend communication tests have been rewritten to support the whole data path from one node to another with a simple writer-reader application.

Metatada operation on files are not yet supported on the memory backend and will be implemented later on.

---------

Co-authored-by: Marco Edoardo Santimaria <marcoedoardo.santimaria@unito.it>