Refactor of project structure
This commit refactors the project structure, moving from the whole
project being inside the src directory, to a common namespace for the
whole components of the CAPIO middleware.

The new namespace (capio) contains the server, posix and common code,
also moving the common/capio code up one level to the common directory.