Ported from CAPIO-V2 CapioClEngine
This commits ports, but not yet integrates the CapioCLEngine Class. This
class is reponsible of replacing the olt logations and glob_locations
global variable, to provide a unified place to handle the logic of the
CAPIO-CL coordination language. This commits also ports two extra
components: the CapioGlobalConfiguration class, which in CAPIO-V2 stores
the global configurations for the currently running server instance, as
well as the JsonParser class, which encapsulates the logic to parse che
CAPIO-CL json configuration file.
This commit has header files and source files split across two different
folder.