Improved Error-Handling and Termination Path (#192)
Previously, the CAPIO POSIX library was terminating immediately
using `exit(EXIT_FAILURE)` upon encountering an error. This
behavior has now been replaced with an exception-based mechanism.

When an error occurs, an exception is thrown instead of exiting directly.
The exception is caught in the `hook()` function, where additional
diagnostic information is printed, such as a dump of the currently
executing syscall. The exception is then rethrown, so that the library still
fails as expected.

This approach provides richer debugging information, especially when
the library is built in `Release` mode.

---------

Co-authored-by: Marco Edoardo Santimaria <marcoedoardo.santimaria@unito.it>